rename package to code_switching
